refactor: split components in diferent files
